Migrate CrashReporting Module
