+++ ios/Classes/CrashReportingApiImpl.h
@@ -1,0 +1,2 @@
+@interface CrashReportingApiImpl : NSObject <CrashReportingApi>
+@end
