+++ pigeons/crash_reporting.api.dart
@@ -1,0 +1,7 @@
+import 'package:pigeon/pigeon.dart';
+
+@HostApi()
+abstract class CrashReportingApi {
+  void setEnabled(bool isEnabled);
+  void send(String jsonCrash, bool isHandled);
+}
